Move lib/haml/plugin.rb to proper path
considering its class name.
