+++ lib/haml/plugin.rb
