+++ lib/haml/template.rb
@@ -36,4 +36,4 @@ end
 Haml::Template.options[:ugly] = defined?(Rails.env) ? !Rails.env.development? : true
 Haml::Template.options[:escape_html] = true
 
-require 'haml/template/plugin'
+require 'haml/plugin'
